+++ lib/legion/extensions/node/actors/crypt.rb
@@ -5,6 +5,8 @@ module Legion
     module Node
       module Actor
         class Crypt < Legion::Extensions::Actors::Subscription
+          def runner_class = Legion::Extensions::Node::Runners::Node
+
           def delay_start
             2
           end